adds useOnOutsideClick hook; draft notification ui
